+++ main/src/main/java/de/learnlib/alex/core/entities/SymbolAction.java
@@ -91,6 +91,7 @@ import java.io.Serializable;
         @JsonSubTypes.Type(name = "web_executeScript", value = ExecuteScriptAction.class),
         @JsonSubTypes.Type(name = "web_fill", value = FillAction.class),
         @JsonSubTypes.Type(name = "web_goto", value = GotoAction.class),
+        @JsonSubTypes.Type(name = "web_moveMouse", value = MoveMouseAction.class),
         @JsonSubTypes.Type(name = "web_submit", value = SubmitAction.class),
         @JsonSubTypes.Type(name = "web_select", value = SelectAction.class),
         @JsonSubTypes.Type(name = "web_waitForTitle", value = WaitForTitleAction.class),