+++ main/src/main/webapp/tests/unit/entities/actions/restActions/CallRestAction.tests.js
@@ -22,7 +22,8 @@ describe('CallRestAction', () => {
             url: '',
             data: null,
             cookies: {},
-            headers: {}
+            headers: {},
+            credentials: {}
         };
         const action = new CallRestAction({});
         expect(angular.toJson(action)).toEqual(angular.toJson(expectedAction));