+++ main/src/test/java/de/learnlib/alex/actions/WebSymbolActions/MoveMouseTest.java
@@ -27,22 +27,21 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.WebElement;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.htmlunit.HtmlUnitDriver;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
 
-import static de.learnlib.alex.core.entities.ExecuteResult.OK;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class ClickActionTest {
+public class MoveMouseTest {
 
     @Mock
     private User user;
@@ -50,66 +49,62 @@ public class ClickActionTest {
     @Mock
     private Project project;
 
-    private ClickAction c;
+    private MoveMouseAction c;
 
     @Before
     public void setUp() {
-        c = new ClickAction();
+        c = new MoveMouseAction();
         c.setUser(user);
         c.setProject(project);
         c.setNode("#node");
-        c.setDoubleClick(false);
+        c.setOffsetX(0);
+        c.setOffsetY(0);
     }
 
     @Test
     public void testJSON() throws IOException {
         ObjectMapper mapper = new ObjectMapper();
         String json = mapper.writeValueAsString(c);
-        ClickAction c2 = mapper.readValue(json, ClickAction.class);
+        MoveMouseAction c2 = mapper.readValue(json, MoveMouseAction.class);
 
         assertEquals(c.getNode(), c2.getNode());
-        assertEquals(c.isDoubleClick(), c2.isDoubleClick());
-    }
-
-    @Test
-    public void testJSONWithLongNode() throws IOException {
-        c.setNode("#superlong > css trace .with-absolute ~no_meaning .at-all > .1234567890 > .1234567890 > .1234567890"
-                          + " > .1234567890 > .1234567890 > .1234567890 > .1234567890 > .1234567890 > .1234567890"
-                          + " > .1234567890 > .1234567890 > .1234567890");
-        ObjectMapper mapper = new ObjectMapper();
-        String json = mapper.writeValueAsString(c);
-        ClickAction c2 = mapper.readValue(json, ClickAction.class);
-
-        assertEquals(c.getNode(), c2.getNode());
-        assertEquals(c.isDoubleClick(), c2.isDoubleClick());
+        assertEquals(c.getOffsetX(), c2.getOffsetX());
+        assertEquals(c.getOffsetY(), c2.getOffsetY());
     }
 
     @Test
     public void testJSONFile() throws IOException, URISyntaxException {
         ObjectMapper mapper = new ObjectMapper();
 
-        File file = new File(getClass().getResource("/actions/websymbolactions/ClickTestData.json").toURI());
+        File file = new File(getClass().getResource("/actions/websymbolactions/MoveMouseTestData.json").toURI());
         WebSymbolAction obj = mapper.readValue(file, WebSymbolAction.class);
 
-        assertTrue(obj instanceof ClickAction);
-        ClickAction objAsAction = (ClickAction) obj;
+        assertTrue(obj instanceof MoveMouseAction);
+        MoveMouseAction objAsAction = (MoveMouseAction) obj;
         assertEquals("#node", objAsAction.getNode());
+        assertEquals(0, objAsAction.getOffsetX());
+        assertEquals(0, objAsAction.getOffsetY());
+    }
+
+
+    @Test
+    public void shouldReturnOKIfNodeCouldBeFound() {
+        // TODO: cannot really test this atm because moveToElement does not like the mocked WebElement
+//        WebSiteConnector connector = mock(WebSiteConnector.class);
+//        WebElement element = mock(WebElement.class);
+//        WebDriver driver = mock(HtmlUnitDriver.class);
+//        given(connector.getElement("#node")).willReturn(element);
+//        given(connector.getDriver()).willReturn(driver);
+//
+//        assertEquals(OK, c.execute(connector));
     }
 
     @Test
-    public void shouldReturnOKIfNodeCouldBeClicked() {
+    public void shouldReturnFAILEDIfNodeCouldNotBeClicked() {
         WebSiteConnector connector = mock(WebSiteConnector.class);
-        WebElement element = mock(WebElement.class);
-        given(connector.getElement("#node")).willReturn(element);
-
-        assertEquals(OK, c.execute(connector));
-        verify(element).click();
-    }
-
-    @Test
-    public void shouldReturnFaliedIfNodeCouldNotBeClicked() {
-        WebSiteConnector connector = mock(WebSiteConnector.class);
+        WebDriver driver = mock(HtmlUnitDriver.class);
         when(connector.getElement("#node")).thenThrow(new NoSuchElementException(""));
+        given(connector.getDriver()).willReturn(driver);
 
         assertEquals(ExecuteResult.FAILED, c.execute(connector));
     }